requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
redis[hiredis]==5.0.1
openai==1.3.5
pyyaml==6.0.1
httpx==0.25.2
python-dotenv==1.0.0
watchfiles==0.21.0
requests==2.31.0

